PEZY-461: Build sticky navigation bar component
