+++ frontend/src/components/NavBar.css
@@ -1,0 +1,267 @@
+.home-police__header {
+  position: sticky;
+  top: 0;
+  z-index: 50;
+  width: 100vw;
+  max-width: 100vw;
+  left: 50%;
+  margin-left: -50vw;
+}
+
+.home-police__container {
+  width: 100%;
+  max-width: none;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+.home-police__topbar {
+  background: linear-gradient(135deg, #0284c7 0%, #0369a1 100%);
+  color: #fff;
+}
+
+.home-police__brand-row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  min-height: 4.5rem;
+  padding: 0.75rem 1.5rem;
+}
+
+.home-police__brand {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.home-police__brand-logo {
+  width: 3.5rem;
+  height: 3.5rem;
+  object-fit: contain;
+  flex-shrink: 0;
+}
+
+.home-police__brand h1 {
+  margin: 0;
+  color: #ffffff;
+  font-size: clamp(1.4rem, 3vw, 2rem);
+  letter-spacing: 0.08em;
+  font-weight: 700;
+  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.home-police__brand-flag {
+  width: 3rem;
+  height: 1.5rem;
+  object-fit: cover;
+  border: 1px solid rgba(255, 255, 255, 0.6);
+  border-radius: 2px;
+  flex-shrink: 0;
+}
+
+.home-police__lang {
+  color: rgba(255, 255, 255, 0.9);
+  font-size: 0.8rem;
+  font-weight: 600;
+  letter-spacing: 0.05em;
+  text-transform: uppercase;
+}
+
+.home-police__menu-row {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr) 12rem;
+  gap: 1rem;
+  background: linear-gradient(135deg, #0284c7 0%, #0369a1 100%);
+  border-top: 1px solid rgba(255, 255, 255, 0.15);
+  padding: 0.5rem 1.5rem;
+  position: relative;
+}
+
+.home-police__menu {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 0.2rem;
+}
+
+.home-police__menu-item {
+  border: 0;
+  background: transparent;
+  color: rgba(255, 255, 255, 0.95);
+  min-height: 2rem;
+  padding: 0.4rem 0.9rem;
+  font-size: 0.825rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 200ms ease;
+  border-radius: 3px;
+}
+
+.home-police__menu-item:hover {
+  background: rgba(255, 255, 255, 0.15);
+  color: #fff;
+}
+
+.home-police__menu-item.is-active {
+  background: rgba(255, 255, 255, 0.25);
+  color: #fff;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+}
+
+.home-police__search {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  border: 1px solid rgba(255, 255, 255, 0.3);
+  background: rgba(255, 255, 255, 0.95);
+  padding: 0.35rem 0.7rem;
+  border-radius: 3px;
+}
+
+.home-police__search input {
+  width: 100%;
+  border: 0;
+  background: transparent;
+  height: 1.8rem;
+  outline: none;
+  color: #4b5563;
+  font-size: 0.85rem;
+}
+
+.home-police__search input::placeholder {
+  color: #9ca3af;
+}
+
+.home-police__menu-toggle {
+  display: none;
+  align-items: center;
+  gap: 0.7rem;
+  width: fit-content;
+  align-self: flex-start;
+  min-height: 2.9rem;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 4px;
+  padding: 0.8rem 1rem;
+  background: rgba(255, 255, 255, 0.22);
+  color: #fff;
+  cursor: pointer;
+  box-shadow: 0 6px 16px rgba(3, 105, 161, 0.28);
+}
+
+.home-police__menu-toggle-icon {
+  display: inline-flex;
+  flex-direction: column;
+  gap: 0.3rem;
+  width: 1.1rem;
+}
+
+.home-police__menu-toggle-icon span {
+  display: block;
+  width: 100%;
+  height: 2px;
+  border-radius: 999px;
+  background: currentColor;
+}
+
+.home-police__menu-toggle-text {
+  font-size: 0.9rem;
+  font-weight: 700;
+  letter-spacing: 0.06em;
+}
+
+.home-police__menu-toggle:hover {
+  background: rgba(255, 255, 255, 0.3);
+  box-shadow: 0 8px 20px rgba(3, 105, 161, 0.34);
+}
+
+.home-police__menu-toggle:active {
+  transform: translateY(1px);
+}
+
+.home-police__mobile-menu {
+  display: none;
+}
+
+@media (max-width: 980px) {
+  .home-police__menu-row {
+    grid-template-columns: clamp(7rem, 24vw, 8.5rem) minmax(0, 1fr);
+    gap: 0.6rem;
+    padding: 0.7rem 1rem 0.85rem;
+    align-items: center;
+  }
+
+  .home-police__menu-toggle {
+    display: flex;
+    width: 100%;
+    min-height: 2.75rem;
+    padding: 0.6rem 0.75rem;
+    grid-column: 1;
+    grid-row: 1;
+  }
+
+  .home-police__menu {
+    display: none;
+    grid-column: 1 / -1;
+    grid-row: 2;
+    width: 100%;
+    flex-direction: column;
+    align-items: stretch;
+    gap: 0;
+  }
+
+  .home-police__menu.is-open {
+    display: flex;
+  }
+
+  .home-police__menu-item {
+    width: 100%;
+    justify-content: flex-start;
+    padding: 0.75rem 0.9rem;
+    text-align: left;
+    border-radius: 0;
+  }
+
+  .home-police__search {
+    grid-column: 2;
+    grid-row: 1;
+    width: 100%;
+    min-width: 0;
+    min-height: 2.75rem;
+    padding: 0.35rem 0.65rem;
+    box-sizing: border-box;
+  }
+
+  .home-police__search input {
+    height: 2rem;
+    font-size: 0.9rem;
+  }
+}
+
+@media (max-width: 680px) {
+  .home-police__menu-row {
+    grid-template-columns: clamp(7.25rem, 34vw, 8.75rem) minmax(0, 1fr);
+  }
+
+  .home-police__brand-logo {
+    width: 3.05rem;
+    height: 3.05rem;
+  }
+
+  .home-police__brand-flag {
+    width: 1.85rem;
+    height: 1.2rem;
+  }
+
+  .home-police__brand h1 {
+    letter-spacing: 0.03em;
+  }
+
+  .home-police__search {
+    min-height: 2.75rem;
+  }
+
+  .home-police__search input {
+    height: 2rem;
+    font-size: 0.9rem;
+  }
+}
