+++ frontend/src/pages/Home.css
@@ -10,6 +10,16 @@
   background: #f8fafc;
 }
 
+.home-police__header {
+  position: sticky;
+  top: 0;
+  z-index: 50;
+  width: 100vw;
+  max-width: 100vw;
+  left: 50%;
+  margin-left: -50vw;
+}
+
 .home-police__container {
   width: 100%;
   max-width: none;
