+++ frontend/src/pages/Home.tsx
@@ -1,23 +1,11 @@
 import { useEffect, useState } from 'react'
 import type { CSSProperties } from 'react'
-import { useNavigate } from 'react-router-dom'
-import pLogo from '../assets/plogo.png'
 import slideOne from '../assets/slider/slide-1.png'
 import slideTwo from '../assets/slider/slide-2.png'
 import slideThree from '../assets/slider/slide-3.png'
 import serviceBg from '../assets/service-bg.jpg'
+import NavBar from '../components/NavBar'
 import './Home.css'
-
-const primaryMenu: { label: string; path?: string }[] = [
-  { label: 'Home', path: '/' },
-  { label: 'About Us' },
-  { label: 'Fine Pay', path: '/fine-pay' },
-  { label: 'Criminal Records', path: '/criminal-records' },
-  { label: 'Division' },
-  { label: 'Downloads' },
-  { label: 'Library' },
-  { label: 'Survey' },
-]
 
 const heroSlides = [
   {
@@ -59,8 +47,6 @@ const servicesData = [
 
 export default function Home() {
   const [activeIndex, setActiveIndex] = useState(0)
-  const [menuOpen, setMenuOpen] = useState(false)
-  const navigate = useNavigate()
 
   useEffect(() => {
     const timer = window.setInterval(() => {
@@ -80,64 +66,9 @@ export default function Home() {
     setActiveIndex(previous => (previous + 1) % heroSlides.length)
   }
 
-  const toggleMenu = () => {
-    setMenuOpen(previous => !previous)
-  }
-
-  const handleMenuItemClick = (path?: string) => {
-    setMenuOpen(false)
-    if (path) {
-      navigate(path)
-    }
-  }
-
   return (
     <section className="home-police">
-      <div className="home-police__topbar">
-        <div className="home-police__container home-police__brand-row">
-          <div className="home-police__brand">
-            <img src={pLogo} alt="Sri Lanka Police logo" className="home-police__brand-logo" />
-            <h1>SRI LANKA POLICE</h1>
-            {/* <img src={sriLankaFlag} alt="Flag of Sri Lanka" className="home-police__brand-flag" /> */}
-          </div>
-          {/* <span className="home-police__lang">ENGLISH</span> */}
-        </div>
-      </div>
-
-      <div className="home-police__container home-police__menu-row">
-        <button
-          type="button"
-          className="home-police__menu-toggle"
-          onClick={toggleMenu}
-          aria-label="Toggle menu"
-          aria-expanded={menuOpen}
-        >
-          <span className="home-police__menu-toggle-icon" aria-hidden="true">
-            <span />
-            <span />
-            <span />
-          </span>
-          <span className="home-police__menu-toggle-text">MENU</span>
-        </button>
-
-        <nav className={`home-police__menu${menuOpen ? ' is-open' : ''}`} aria-label="Primary">
-          {primaryMenu.map(item => (
-            <button
-              key={item.label}
-              type="button"
-              className={`home-police__menu-item${item.label === 'Home' ? ' is-active' : ''}`}
-              onClick={() => handleMenuItemClick(item.path)}
-            >
-              {item.label}
-            </button>
-          ))}
-        </nav>
-
-        <div className="home-police__search">
-          <span aria-hidden="true">&#128269;</span>
-          <input type="text" value="" readOnly placeholder="Search" aria-label="Search" />
-        </div>
-      </div>
+      <NavBar activeLabel="Home" />
 
       <div className="home-police__container home-police__news-row">
         <span className="home-police__news-tag">News &#187;</span>
